Make deviceId an int everywhere

+++ service/src/main/java/org/whispersystems/signalservice/internal/push/PushServiceSocket.java
@@ -733,7 +733,7 @@ public class PushServiceSocket implements Closeable {
     return JsonUtil.fromJsonResponse(response, RestoreMethodBody.class);
   }
 
-  public void removeDevice(long deviceId) throws IOException {
+  public void removeDevice(int deviceId) throws IOException {
     makeServiceRequest(String.format(DEVICE_PATH, String.valueOf(deviceId)), "DELETE", null);
   }
